backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
playwright==1.48.0
boto3==1.35.0
python-dotenv==1.0.1
httpx==0.27.2
pydantic==2.9.0
reportlab==4.2.2
python-multipart==0.0.9
